+++ src/realm/parser/driver.cpp
@@ -1407,9 +1407,12 @@ std::unique_ptr<Subexpr> ConstantNode::visit(ParserDriver* drv, DataType hint)
                 case type_Float:
                     ret = std::make_unique<Value<float>>(float(double_val));
                     break;
-                case type_Decimal:
-                    ret = std::make_unique<Value<Decimal128>>(Decimal128(text));
+                case type_Decimal: {
+                    // If not argument, try decode again to get full precision
+                    Decimal128 dec = (type == Type::ARG) ? Decimal128(double_val) : Decimal128(text);
+                    ret = std::make_unique<Value<Decimal128>>(dec);
                     break;
+                }
                 case type_Int: {
                     int64_t int_val = int64_t(double_val);
                     // Only return an integer if it precisely represents val